resin-init-flasher: scan the original image when repartitioning for LUKS
In the current state flasher dd's the partition table and looks for partition
metadata on the new device, which is pretty fragile, because the partition
table may get corrupted. We also want to remove this trick and repartition
properly.

With this patch flasher scans for the original image attached to a loop device
for the partition metadata, which is r/o and assumed to be valid from build.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>