common: add crun to packagegroup-resin
Add crun to packagegroup-resin, ensuring it's installed in all images.
With this patch, crun can be used in place of the default golang runc
runtime.

This runtime is smaller, and allows for faster container startup and
lower resource usage compared to runc.

The runtime can either be specified on the command line:

$ balena run --runtime=crun hello-world

Or as the default runtime in the engine's daemon.json config:

$ cat << EOF > /etc/docker/daemon.json
{
  "runtimes": {
    "crun": {"path": "/usr/bin/crun"}
  },
  "default-runtime": "crun"
}
EOF

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>